code/buffer/functions/proxy/requirements.txt
requests==2.5.0
python-dotenv==0.6.2
boto3==1.4.4
